fix: split reconciliation status budget test
Move the reconciliation status diagnostics test into a focused module.

Keep Rust line budgets and backend shard coverage intact.
